Fixed Calculating the TF-IDF value - Math.log10 needed casting values to double. Test corrected

+++ src/main/java/com/czajor/simplesearchengine/service/CalculateTfIdfService.java
@@ -8,7 +8,7 @@ public class CalculateTfIdfService {
     }
 
     public double calculateIdf(int allDocuments, int docsContainingWord) {
-        return (double) Math.log10(allDocuments / docsContainingWord);
+        return Math.log10((double) allDocuments / (double) docsContainingWord);
     }
 
     public double calculateTfIdf(double tf, double idf) {
